Adapted to the v4 version and created a new. go file prasing_ AOF ，further optimized the code structure and tested the cluster.
